[svn-r1341] Changes since 19990611 ----------------------
./doc/html/H5.format.html
	Added documentation for opaque data types (bitfield types were
	already documented but they were out of order).

./src/H5E.c
	Fixed a bug with glibc2 on linux systems where `stdout' is an
	extern and can't be used to initialize static data.

./src/H5T.c
./src/H5Tpkg.h
./src/H5Tpublic.h
./src/H5detect.c
	Removed the `_T' from the new C9x types I just added so the
	names are consistent with existing types. Besides, the fact
	that something is a datatype is obvious because it starts with
	H5T_NATIVE_.

./tools/h5ls.c
	Added the new C9x data types. H5ls prints one of these types
	only if it doesn't match one of the builtin C types.

	Prints the OID for shared data types.

	Fixed a formatting bug with symbolic links which was
	introduced a few changes ago.

	The commandline has been changed so that objects from multiple
	files can be listed with a single command. Instead of
	specifying a file name and an optional list of objects, each
	thing to print is a file name and object concatenated. H5ls
	figures out how to devide the name into a file name and object
	name even when the file name part doesn't correspond to an
	actual Unix file.

	   Old syntax: h5ls [OPTIONS] FILE [OBJECTS]
	   New syntax: h5ls [OPTIONS] FILE[/OBJECT] [FILE[/OBJECT]]...

	   Example ({X,Y} is expanded by the shell)

	   Old command: h5ls -d ../test/x.data dir1 dir2
	   New command: h5ls -d ../test/x.data/{dir1,dir2}

	The filename is printed as part of the object name when full
	names are requested (--full or --recursive). If people really
	don't like this they can undefine a constant at the top of
	h5ls.

	Errors from the hdf5 library are turned off.

	Commandline switches of the form `--width 80' are accepted in
	addition to `--width=80'. This is more symmetric with
	single-letter switches that take two forms: `-w 80' and
	`-w80'.

./src/H5D.c
	Added tracing instrumentation for H5Dvlen_reclaim().

./src/H5private.h
	Added casts to int for the isalpha() et al macros to shut up
	solaris warnings about char subscripts.

+++ src/H5private.h
@@ -545,18 +545,18 @@ __DLL__ int HDfprintf (FILE *stream, const char *fmt, ...);
 #define HDgets(S)		gets(S)
 #define HDgetuid()		getuid()
 #define HDgmtime(T)		gmtime(T)
-#define HDisalnum(C)		isalnum(C)
-#define HDisalpha(C)		isalpha(C)
+#define HDisalnum(C)		isalnum((int)(C)) /*cast for solaris warning*/
+#define HDisalpha(C)		isalpha((int)(C)) /*cast for solaris warning*/
 #define HDisatty(F)		isatty(F)
-#define HDiscntrl(C)		iscntrl(C)
-#define HDisdigit(C)		isdigit(C)
-#define HDisgraph(C)		isgraph(C)
-#define HDislower(C)		islower(C)
-#define HDisprint(C)		isprint(C)
-#define HDispunct(C)		ispunct(C)
-#define HDisspace(C)		isspace(C)
-#define HDisupper(C)		isupper(C)
-#define HDisxdigit(C)		isxdigit(C)
+#define HDiscntrl(C)		iscntrl((int)(C)) /*cast for solaris warning*/
+#define HDisdigit(C)		isdigit((int)(C)) /*cast for solaris warning*/
+#define HDisgraph(C)		isgraph((int)(C)) /*cast for solaris warning*/
+#define HDislower(C)		islower((int)(C)) /*cast for solaris warning*/
+#define HDisprint(C)		isprint((int)(C)) /*cast for solaris warning*/
+#define HDispunct(C)		ispunct((int)(C)) /*cast for solaris warning*/
+#define HDisspace(C)		isspace((int)(C)) /*cast for solaris warning*/
+#define HDisupper(C)		isupper((int)(C)) /*cast for solaris warning*/
+#define HDisxdigit(C)		isxdigit((int)(C)) /*cast for solaris warning*/
 #define HDkill(P,S)		kill(P,S)
 #define HDlabs(X)		labs(X)
 #define HDldexp(X,N)		ldexp(X,N)
